+++ expyriment/control/_experiment_control.py
@@ -140,7 +140,9 @@ def start(experiment=None, auto_create_subject_id=None, subject_id=None,
                     message="Subject Number:",
                     position=position,
                     message_colour=misc.constants.C_EXPYRIMENT_PURPLE,
+                    message_text_size=int(experiment.text_size * 1.2),
                     user_text_colour=misc.constants.C_EXPYRIMENT_ORANGE,
+                    user_text_size=int(experiment.text_size),
                     background_colour=(0, 0, 0),
                     frame_colour=(70, 70, 70),
                     ascii_filter=misc.constants.K_ALL_DIGITS)
@@ -179,6 +181,7 @@ def start(experiment=None, auto_create_subject_id=None, subject_id=None,
     number = defaults.initialize_delay - int(experiment.clock.time // 1000)
     if number > 0:
         text = stimuli.TextLine("Initializing, please wait...",
+                                text_size=int(experiment.text_size * 1.2),
                                 text_colour=(160, 70, 250),
                                 position=(0, 0))
         stimuli._stimulus.Stimulus._id_counter -= 1
@@ -189,7 +192,7 @@ def start(experiment=None, auto_create_subject_id=None, subject_id=None,
     while number > 0:
         counter = stimuli.TextLine(
             "{num:02d}".format(num=number),
-            text_size= int(default_textline_size * 0.8),
+            text_size= int(experiment.text_size * 0.9),
             text_font='FreeMono',
             text_bold=True,
             text_colour=misc.constants.C_EXPYRIMENT_ORANGE,
@@ -207,7 +210,8 @@ def start(experiment=None, auto_create_subject_id=None, subject_id=None,
     position = (0, 0)
     if not skip_ready_screen:
         stimuli.TextLine("Ready", position=position,
-                     text_colour=misc.constants.C_EXPYRIMENT_ORANGE).present()
+                         text_size=int(experiment.text_size * 1.2),
+                         text_colour=misc.constants.C_EXPYRIMENT_ORANGE).present()
         stimuli._stimulus.Stimulus._id_counter -= 1
         if android is None:
             experiment.keyboard.wait()
@@ -241,7 +245,8 @@ def pause():
     else:
         position = (0, 0)
     stimuli.TextLine("Paused", position=position,
-                     text_colour=misc.constants.C_EXPYRIMENT_ORANGE).present()
+                     text_colour=misc.constants.C_EXPYRIMENT_ORANGE,
+                     text_size=int(_internals.active_exp.text_size * 1.2)).present()
     experiment.set_log_level(old_logging)
     experiment._screen.colour = screen_colour
     stimuli._stimulus.Stimulus._id_counter -= 1
@@ -297,7 +302,8 @@ def end(goodbye_text=None, goodbye_delay=None, confirmation=False,
         else:
             position = (0, 0)
         stimuli.TextLine("Quitting Experiment? (y/n)", position=position,
-                         text_colour=misc.constants.C_EXPYRIMENT_ORANGE).present()
+                         text_colour=misc.constants.C_EXPYRIMENT_ORANGE,
+                         text_size=int(_internals.active_exp.text_size * 1.2)).present()
         stimuli._stimulus.Stimulus._id_counter -= 1
         char = Keyboard().wait_char(["y", "n"], process_control_events=False)
         if char[0] == "n":
@@ -317,7 +323,7 @@ def end(goodbye_text=None, goodbye_delay=None, confirmation=False,
         fast_quit = defaults.fast_quit
     if fast_quit and experiment.is_started:
         if experiment.screen.window_mode:
-            if expyriment.screen.window_no_frame:
+            if experiment.screen.no_frame:
                 pygame.display.set_mode(experiment.screen._window_size, pygame.NOFRAME)
             else:
                 pygame.display.set_mode(experiment.screen._window_size)
@@ -326,7 +332,8 @@ def end(goodbye_text=None, goodbye_delay=None, confirmation=False,
     try:
         experiment._screen.colour = [0, 0, 0]
         stimuli.TextLine(goodbye_text, position=(0, 0),
-                         text_colour=misc.constants.C_EXPYRIMENT_PURPLE).present()
+                         text_colour=misc.constants.C_EXPYRIMENT_PURPLE,
+                         text_size=int(_internals.active_exp.text_size * 1.2)).present()
         stimuli._stimulus.Stimulus._id_counter -= 1
     except:
         pass
@@ -413,7 +420,7 @@ fullscreen.""")
                                 open_gl=defaults.open_gl,
                                 window_mode=defaults.window_mode,
                                 window_size=defaults.window_size,
-                                no_frame = defaults.windows_no_frame)
+                                no_frame=defaults.window_no_frame)
     # Hack for IDLE: quit pygame and call atexit functions when crashing
     if misc.is_idle_running() and sys.argv[0] != "":
         try:
@@ -442,12 +449,11 @@ fullscreen.""")
         experiment._events = None
     experiment._keyboard = Keyboard()
     experiment._mouse = Mouse(show_cursor=False)
-    default_textline_size = stimuli.TextLine(text="").text_size
     logo = stimuli.Picture(misc.constants.EXPYRIMENT_LOGO_FILE,
                            position=(0, 100))
     logo.scale((0.7, 0.7))
     text = stimuli.TextLine("Version {0}".format(get_version()),
-                            text_size= int(default_textline_size * 0.8),
+                            text_size= experiment.text_size,
                             text_colour=misc.constants.C_EXPYRIMENT_PURPLE,
                             background_colour=(0, 0, 0),
                             position=(0, -5))
@@ -462,7 +468,7 @@ fullscreen.""")
             txt += ", {0}".format(
                         misc.module_hashes_as_string())
         text2 = stimuli.TextLine(txt,
-            text_size= int(default_textline_size * 0.5),
+            text_size= int(experiment.text_size * 0.7),
             text_colour=misc.constants.C_EXPYRIMENT_ORANGE,
             background_colour=(0, 0, 0),
             position=(0, -50))
@@ -505,7 +511,7 @@ fullscreen.""")
                                                 check_for_control_keys=False)
                 if key is not None:
                     break
-    stimuli.TextLine("Preparing experiment...", text_size=24,
+    stimuli.TextLine("Preparing experiment...", text_size=int(experiment.text_size * 1.2),
                      text_colour=misc.constants.C_EXPYRIMENT_PURPLE).present()
     experiment._screen.colour = experiment.background_colour
     experiment.set_log_level(old_logging)